Changing cookie to session for logging out
NOTE: all of the navigation buttons do not go to the link they are supposed to go to.
I made a form tag and put a logout button on it to test it out. This is the only navigation button that actually works. We need to change all the other buttons in order to make them work, too.